Restructure architecture so classes are associated with feature.
