+++ app/src/main/java/pitcherseye/pitcherseye/Tagging/ChangePitcherFragment.java
@@ -5,7 +5,7 @@
  session using the PitcherStats object to Firebase.
  */
 
-package pitcherseye.pitcherseye.Fragments;
+package pitcherseye.pitcherseye.Tagging;
 
 import android.app.DialogFragment;
 import android.content.Context;
@@ -32,7 +32,7 @@ import java.util.List;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
-import pitcherseye.pitcherseye.Activities.TaggingActivity;
+import pitcherseye.pitcherseye.Tagging.TaggingActivity;
 import pitcherseye.pitcherseye.R;
 
 public class ChangePitcherFragment extends DialogFragment {
