+++ app/src/main/res/layout/activity_tagging.xml
@@ -5,7 +5,7 @@
                                              android:layout_width="match_parent"
                                              android:layout_height="match_parent"
                                              android:background="@color/colorTaggingBackground"
-                                             tools:context="pitcherseye.pitcherseye.Activities.TaggingActivity">
+                                             tools:context="pitcherseye.pitcherseye.Tagging.TaggingActivity">
 
 
     <!-- Header -->
