+++ app/src/main/res/layout/fragment_change_pitcher.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorTaggingBackground"
-    tools:context="pitcherseye.pitcherseye.Activities.TaggingActivity"
+    tools:context="pitcherseye.pitcherseye.Tagging.TaggingActivity"
     android:focusableInTouchMode="true" >
 
     <View
